♻️ 5G5506 code: deduplicate CLI benchmark helpers
