+++ .agentplane/tasks/202605091753-5G5506/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605091753-5G5506/benchmark-helpers",
+  "created_at": "2026-05-09T18:22:55.946Z",
+  "head_sha": "d25d4ae5fb957961719455f3b4ade92166a0dd27",
+  "last_verified_at": "2026-05-09T18:28:03.684Z",
+  "last_verified_sha": "d25d4ae5fb957961719455f3b4ade92166a0dd27",
+  "schema_version": 1,
+  "task_id": "202605091753-5G5506",
+  "updated_at": "2026-05-09T18:22:55.946Z",
+  "verify": {
+    "status": "pass"
+  }
+}
