+++ .agentplane/tasks/202605091753-5G5506/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-09T18:22:55.946Z
+
+## Task
+
+- Task: `202605091753-5G5506`
+- Title: Deduplicate CLI benchmark script helpers
+- Status: DOING
+- Branch: `task/202605091753-5G5506/benchmark-helpers`
+- Canonical task record: `.agentplane/tasks/202605091753-5G5506/README.md`
+
+## Verification
+
+- State: ok
+- Note: Verified: extracted shared benchmark helper module; help paths, walltime smoke, perf smoke, Prettier, typecheck, clone:report, and clone:check passed. Clone metrics improved from 1546 duplicated lines / 16193 tokens after the previous task to 1465 duplicated lines / 15457 tokens, with clone count 88 -> 85.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-09T18:22:55.946Z
+- Branch: task/202605091753-5G5506/benchmark-helpers
+- Head: d25d4ae5fb95
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
